Refactor out `lovatAPI.getTournaments()`
