+++ lib/reusable/lovat_api/lovat_api.dart
@@ -133,37 +133,7 @@ class LovatAPI {
     });
   }
 
-  // Specific endpoints
-  Future<PartialTournamentList> getTournaments({
-    int? take,
-    int? skip,
-    String filter = '',
-  }) async {
-    final response = await get(
-      '/v1/manager/tournaments',
-      query: {
-        if (take != null) 'take': take.toString(),
-        if (skip != null) 'skip': skip.toString(),
-        'filter': filter,
-      },
-    );
-
-    if (response?.statusCode != 200) {
-      throw Exception('Failed to get tournaments');
-    }
-
-    final json = jsonDecode(response!.body) as Map<String, dynamic>;
-    final tournamentJson = json['tournaments'] as List<dynamic>;
-
-    final tournaments =
-        tournamentJson.map((e) => Tournament.fromJson(e)).toList();
-
-    return PartialTournamentList(
-      tournaments: tournaments,
-      total: json['count'],
-    );
-  }
-
+  // MARK: Endpoints
   Future<void> setUsername(String username) async {
     final response = await post(
       '/v1/manager/onboarding/username',
@@ -1184,16 +1154,6 @@ extension CSVExportModeExtension on CSVExportMode {
         return 'Each row contains data collected by one scouter about a specific team\'s performance during a match.';
     }
   }
-}
-
-class PartialTournamentList {
-  const PartialTournamentList({
-    required this.tournaments,
-    required this.total,
-  });
-
-  final List<Tournament> tournaments;
-  final int total;
 }
 
 enum RegistrationStatus {
